requirements.txt
requests
pillow
matplotlib
torch
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
pandas